Add sub container to fix formatting
Sub container contains each bar of the array to be sorted. Intead of each bar being its own container on the same level, the entire collection of bars is on the same level as the buttons. This allows the bars to be side by side horizontally while the buttons are vertical.